[touch: 918] feature: CRUD Speaker
add saveImage and deleteImage methods to BaseController
change validation in TypeRequest
add ckeditor

+++ app/Repositories/BaseRepository.php
@@ -5,9 +5,11 @@ namespace App\Repositories;
 use Bosnadev\Repositories\Contracts\RepositoryInterface;
 use Bosnadev\Repositories\Eloquent\Repository;
 use Carbon\Carbon;
+use Mockery\CountValidator\Exception;
 use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
 use Intervention\Image\Facades\Image;
 use Illuminate\Support\Facades\File;
+use Symfony\Component\HttpFoundation\File\Exception\FileException;
 
 class BaseRepository extends Repository implements RepositoryInterface
 {
@@ -71,9 +73,10 @@ class BaseRepository extends Repository implements RepositoryInterface
      */
     public function saveImage($image, $directory = '', $size = ['width' => 800, 'height' => 600])
     {
-        try {
+         try {
+            $directory = str_replace('.', DIRECTORY_SEPARATOR, $directory);
             $imageRealPath = $image->getRealPath();
-            $timestamp = str_replace([' ', ':'], '-', Carbon::now()->toDateTimeString());
+            $timestamp = Carbon::now()->format('Y-m-d-H-i-s');
             $thumbName = $timestamp . '-' . $image->getClientOriginalName();
 
             $img = Image::make($imageRealPath);
@@ -86,8 +89,13 @@ class BaseRepository extends Repository implements RepositoryInterface
                     $constraint->aspectRatio();
                 });
             }
-            $path = 'uploads' . DIRECTORY_SEPARATOR . $directory . DIRECTORY_SEPARATOR . $thumbName;
-            $img->save(public_path() . DIRECTORY_SEPARATOR . $path);
+            $pathToDirectory = DIRECTORY_SEPARATOR . 'uploads' . DIRECTORY_SEPARATOR . $directory . DIRECTORY_SEPARATOR;
+            if (!$this->checkAndMakeDirectory($pathToDirectory)) {
+                throw new FileException(sprintf('Unable to create the "%s" directory', $pathToDirectory));
+            }
+
+            $path = $pathToDirectory . $thumbName;
+            $img->save(public_path() . $path);
 
             return $path;
         } catch (Exception $e) {
@@ -105,6 +113,26 @@ class BaseRepository extends Repository implements RepositoryInterface
     public function deleteImage($path)
     {
         $path = public_path() . DIRECTORY_SEPARATOR . $path;
-        return File::delete($path);
+        if (File::exists($path)) {
+            return File::delete($path);
+        }
+
+        return true;
+    }
+
+    /**
+     * Create directory if it don't exist
+     *
+     * @param string $path
+     *
+     * @return bool
+     */
+    protected function checkAndMakeDirectory($path)
+    {
+        if (!File::exists($path)) {
+            return File::makeDirectory(public_path() . $path, 0775, true);
+        }
+
+        return true;
     }
 }
